+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 ### Added
+- added combined equipment carbon dioxide emissions report
 - added equipment carbon dioxide emissions report
 - added shopfloor carbon dioxide emissions report
 - added store carbon dioxide emissions report
+- added space carbon dioxide emissions report
+- added tenant carbon dioxide emissions report
 ### Changed
 - 
 ### Fixed